In this task, you’ll replace the static list of Pokémon from the previous Exercises with the complete list of Pokémon loaded from the free Pokémon API.
Add polyfills for promises and fetch to your project.
In the “scripts.js” file of your project, remove the array of Pokémon objects and replace it with an empty array.
Add function(s) inside your data repository to load data from an external source.
Add a LoadList() function as a return key that uses fetch to GET the complete list of Pokémon from here: https://pokeapi.co/api/v2/pokemon/
Use the add() function to add each Pokémon from the results to your pokemonList variable. Make sure to set name and detailsUrl as the keys. Use this as a reference to see what the API response looks like.
Add a loadDetails() function, as well. The loadDetails() function should expect a parameter with a Pokémon object as a parameter. loadDetails() should GET the Pokémon details using the URL from the Pokémon object in the parameter.
Once the GET request is complete, use then to return a JSON response.
then, assign some of the details you got from the response to the Pokémon in the pokemonList. Assign at least imgUrl and height.
Make sure both functions LoadList() and loadDetails() are assigned to keys with the same name in the returned object of your pokemonRepository.
Make sure the Pokémon list is only rendered after you’ve gotten all information from the server.
Call the LoadList() function of pokemonRepository.
Then, execute getAll from the pokemonRepository.
forEach Pokémon, call the addListItem() function you wrote in the previous Exercise.
Edit your showDetails() function to load the Pokémon details from the API instead of loading static data:
Inside the showDetails() function, call the loadDetails() function from above. Pass as parameter the Pokémon object.
Log the result in the console (you’ll display it in the interface in the next Exercise).
Make sure both functions LoadList() and loadDetails() are assigned to keys with the same name in the returned object of your pokemonRepository.
You should now see a list displaying all Pokémon. When you click one, it might take a short moment to load before you see a console.log() with the returned Pokémon object.
Commit the changes in Git and submit the link to your GitHub repository here. Feel free to share additional thoughts or ask questions on your submission page.
Bonus Task

You could also opt to implement the following functionality:

Display a loading message while data is being loaded. To do this, implement a showLoadingMessage() and hideLoadingMessage() function, which appends/removes a message to your page. Hint #1: showLoadingMessage() should be the first thing executed inside both the LoadList() and loadDetails() functions. Hint #2: hideLoadingMessage() should be executed in the then and catch blocks of the fetch code of both LoadList() and loadDetails(). This makes sense because once you’re inside then or catch, it means you’ve received the response from the fetch code and can hide the loading message.